+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * backend: remove panic testing route 
 * backend: cleanup some todos around event validation 
+* frontend: merge multiple bool flags into bitflag (optimization and clippy cleanuo)
 
 ## [2.3.4] - 2022-08-22